+++ src/openlayers/overlay/HeatMap.js
@@ -439,7 +439,7 @@ export class HeatMap extends ol.source.ImageCanvas {
                 return featuresTemp;
             }
         }
-        throw new Error("features 类型不符，请检查。");
+        throw new Error("Features's type does not match, please check.");
     }
 
 }